Move definition of QPDF::JSONReactor into QPDF_json
Allow access to private header files when defining data members.
